# Conflicts:
#	frontend/src/app/AC/products.js
#	frontend/src/tpl/layouts/default.nunj
#	kadena2.0/CMS/css/main.min.css
#	kadena2.0/CMS/gfx/svg/sprites/react.svg
#	kadena2.0/CMS/js/app.min.js
#	kadena2.0/CMS/js/common.min.js

+++ kadena2.0/Kadena2.0.MicroserviceClients/Contracts/ITaxEstimationService.cs
@@ -6,6 +6,6 @@ namespace Kadena2.MicroserviceClients.Contracts
 {
     public interface ITaxEstimationServiceClient
     {
-        Task<BaseResponseDto<double>> CalculateTax(string serviceEndpoint, TaxCalculatorRequestDto request);
+        Task<BaseResponseDto<decimal>> CalculateTax(string serviceEndpoint, TaxCalculatorRequestDto request);
     }
 }
